# docs/openapi_dump.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiskOrbit Risk Scoring & Investigation API",
    "description": "Phase 1 + Phase 2 \u2014 Coordinated Refund Abuse Detection & Relationship Investigation. Phase 1: transaction-level risk scoring. Phase 2: bounded graph investigation and grounded evidence generation.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/risk/score": {
      "post": {
        "summary": "Score Transaction",
        "description": "Score a transaction for coordinated refund abuse risk (Phase 1 baseline).",
        "operationId": "score_transaction_risk_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/risk/investigate": {
      "post": {
        "summary": "Investigate Candidate",
        "description": "Phase 2: Trigger bounded relationship investigation around a candidate.\n\nInput: candidate_id (customer_id or transaction_id), optional as_of_timestamp, max_hops\nOutput: Structured investigation case with evidence, ring, timeline, and grounded narrative.\n\nThis is READ-ONLY. No accounts are blocked or modified.",
        "operationId": "investigate_candidate_risk_investigate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseInvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/risk/cases/{case_id}": {
      "get": {
        "summary": "Get Case",
        "description": "Retrieve a stored investigation case by case_id.",
        "operationId": "get_case_risk_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseInvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/risk/cases/{case_id}/graph": {
      "get": {
        "summary": "Get Case Graph",
        "description": "Return graph statistics for an investigation case.",
        "operationId": "get_case_graph_risk_cases__case_id__graph_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Case Graph Risk Cases  Case Id  Graph Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/risk/cases/{case_id}/evidence": {
      "get": {
        "summary": "Get Case Evidence",
        "description": "Return the grounded evidence bundle for an investigation case.",
        "operationId": "get_case_evidence_risk_cases__case_id__evidence_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Case Evidence Risk Cases  Case Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseAuditTrail": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "scoring_timestamp": {
            "type": "string",
            "title": "Scoring Timestamp"
          },
          "phase1_model_version": {
            "type": "string",
            "title": "Phase1 Model Version",
            "default": "riskorbit-risk-v1"
          },
          "graph_version": {
            "type": "string",
            "title": "Graph Version",
            "default": "graph-v2"
          },
          "investigation_version": {
            "type": "string",
            "title": "Investigation Version",
            "default": "investigation-v2"
          },
          "evidence_version": {
            "type": "string",
            "title": "Evidence Version",
            "default": "evidence-grounded-v2"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "candidate_id",
          "scoring_timestamp",
          "created_at"
        ],
        "title": "CaseAuditTrail"
      },
      "CaseInvestigationRequest": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id",
            "description": "Customer ID or Transaction ID to investigate"
          },
          "as_of_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Timestamp",
            "description": "ISO timestamp for point-in-time scoring"
          },
          "max_hops": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Max Hops",
            "default": 2
          }
        },
        "type": "object",
        "required": [
          "candidate_id"
        ],
        "title": "CaseInvestigationRequest"
      },
      "CaseInvestigationResponse": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "root_entity": {
            "type": "string",
            "title": "Root Entity"
          },
          "candidate_ring_type": {
            "type": "string",
            "title": "Candidate Ring Type",
            "default": "coordinated_refund_abuse"
          },
          "individual_risk_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Individual Risk Score"
          },
          "relationship_risk_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relationship Risk Score"
          },
          "combined_risk_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Combined Risk Score"
          },
          "risk_tier": {
            "type": "string",
            "title": "Risk Tier",
            "description": "low | medium | high | critical"
          },
          "is_candidate_ring": {
            "type": "boolean",
            "title": "Is Candidate Ring"
          },
          "member_accounts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Member Accounts"
          },
          "shared_devices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shared Devices"
          },
          "shared_ips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shared Ips"
          },
          "targeted_merchants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Targeted Merchants"
          },
          "evidence_records": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRecord"
            },
            "type": "array",
            "title": "Evidence Records"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "case_summary": {
            "type": "string",
            "title": "Case Summary"
          },
          "why_flagged": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Why Flagged"
          },
          "uncertainties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainties"
          },
          "missing_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Evidence"
          },
          "graph_statistics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Graph Statistics"
          },
          "audit_trail": {
            "$ref": "#/components/schemas/CaseAuditTrail"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "root_entity",
          "individual_risk_score",
          "relationship_risk_score",
          "combined_risk_score",
          "risk_tier",
          "is_candidate_ring",
          "case_summary",
          "audit_trail"
        ],
        "title": "CaseInvestigationResponse"
      },
...
      "EvidenceCategory": {
        "type": "string",
        "enum": [
          "DEVICE_SHARING",
          "NETWORK_SHARING",
          "INSTRUMENT_LINK",
          "REFUND_COORDINATION",
          "CROSS_MERCHANT",
          "TEMPORAL_BURST",
          "BENIGN_CONFUSER"
        ],
        "title": "EvidenceCategory"
      },
      "EvidenceConfidence": {
        "type": "string",
        "enum": [
          "STRONG",
          "MODERATE",
          "WEAK"
        ],
        "title": "EvidenceConfidence"
      },
      "EvidenceRecord": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id",
            "description": "Unique evidence ID e.g. EVID-0001"
          },
          "category": {
            "$ref": "#/components/schemas/EvidenceCategory"
          },
          "confidence": {
            "$ref": "#/components/schemas/EvidenceConfidence"
          },
          "statement": {
            "type": "string",
            "title": "Statement",
            "description": "Factual evidence claim directly derived from data"
          },
          "supporting_entity_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supporting Entity Ids"
          },
          "supporting_event_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supporting Event Ids"
          },
          "metric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Value"
          },
          "metric_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Description"
          }
        },
        "type": "object",
        "required": [
          "evidence_id",
          "category",
          "confidence",
          "statement"
        ],
        "title": "EvidenceRecord"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScoreRequest": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id",
            "description": "Synthetic customer identifier"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id",
            "description": "Transaction to score"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "transaction_id"
        ],
        "title": "ScoreRequest",
        "description": "Input to POST /risk/score"
      },
      "ScoreResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "risk_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "risk_label": {
            "type": "string",
            "title": "Risk Label",
            "description": "low | medium | high"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "top_signals": {
            "items": {
              "$ref": "#/components/schemas/SignalDetail"
            },
            "type": "array",
            "title": "Top Signals"
          },
          "scored_at": {
            "type": "string",
            "title": "Scored At"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "customer_id",
          "risk_score",
          "risk_label",
          "threshold",
          "model_version",
          "scored_at"
        ],
        "title": "ScoreResponse"
      },
...
      "SignalDetail": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature",
            "description": "Feature name from the canonical feature list"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Feature value for this transaction"
          },
          "importance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Importance",
            "description": "Normalized SHAP importance (0-1) for this feature"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "value",
          "importance"
        ],
        "title": "SignalDetail",
        "description": "One contributing signal in the risk explanation."
      },
...
      "TimelineEvent": {
        "properties": {
          "event_timestamp": {
            "type": "string",
            "title": "Event Timestamp"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "supporting_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supporting Event Id"
          }
        },
        "type": "object",
        "required": [
          "event_timestamp",
          "event_type",
          "entity_id",
          "description"
        ],
        "title": "TimelineEvent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}